requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
click==8.1.3
colorama==0.4.5
Django==4.1.1
djangorestframework==3.14.0
h11==0.13.0
psycopg2==2.9.3
pytz==2022.2.1
sqlparse==0.4.3
tzdata==2022.2
uvicorn==0.18.3
